Rename the folder and update requirements
