+++ requirements.txt
@@ -1,13 +1,2 @@
-aiohttp==3.8.4
-aiosignal==1.3.1
-async-timeout==4.0.2
-attrs==22.2.0
-awesomeversion==22.9.0
-backoff==2.2.1
-cachetools==5.3.0
-charset-normalizer==3.1.0
-frozenlist==1.3.3
-idna==3.4
-multidict==6.0.4
-wled==0.16.0
-yarl==1.8.2
+async-timeout==4.0.3
+wled==0.17.0
